requirements.txt
# annotated-types # 型ヒントの拡張
# anyio # 非同期を簡単にする
# click # コマンドラインアプリ用
# fastapi # フレームワーク
# h11 # HTTPを使うライブラリ
# idna # インターネットドメインを処理
# pip # パッケージ管理
# pydantic # データバリデーション、設定管理
# pydantic_core # pydanticのコア
# sniffio # 現在実行中の非同期ライブラリを識別
# starlette # FastAPIの基盤フレームワーク
# typing_extensions # typingモジュールの拡張
# uvicorn # 高速な非同期特化のサーバ
fastapi==0.110.0
uvicorn[standard]==0.27.1
pytz==2025.2
sqlalchemy==2.0.20
databases==0.9.0
requests==2.32.5
python-multipart==0.0.20
ulid-py==1.1.0
mysql-connector-python==9.5.0
alembic==1.11.1
python-dotenv==1.0.0
pymysql==1.1.2
utils==1.0.2
# Azure Communication Services for Email
azure-communication-email==1.30.0